Estimate_d.c
Here is Estimate_d.c a driver for Numerical Recipes in C routine locate.c. While compiling the driver for locate from the [C] Example Book Second Edition, I noticed the line 'x=-100+200*i/20;' as, to put it bluntly an addition problem. Replacing the line with a loop for x (which I found need not loop) it was immediately apparent the locate routine isolated a certain value immediately, plus one.